requirements.txt
aiogram
beautifulsoup4
curl_cffi
google-generativeai
python-dotenv
